big commit -- lots of electricity documentation work + mystery NG adds..
to be frank, I don't understand why all my NG docs made their way into this commit. I'm not sure of the best path forward, but am sending a commit to not lose my work. I can always restore the last commit to the 24 branch -- that reflects where I want to be at.