New annual land cover files from NLCD
Rolling out new land cover area estimates based on NLCD. In October of 2024, USGS released a new land cover dataset which, among other updates, provided data for interstitial years (based on USGS machine learning methods). This warranted an update on its own, but I've also added some new rules around how we calculate the area of Urban_Tree, Urban_Grassland, and Built-Up.

Previous versions labeled any Developed lands with >0% tree canopy as "Urban_Tree", and the area was recomputed based on the percentage of that tree canopy cover. But what about the residual area? It was unaccounted for.

Now for the cases where we have Developed, Open Space with some tree canopy, the pixel gets assigned "Urban_Tree", area is recomputed based on tree canopy percentage, but then the residual area is labeled "Urban_Grassland" and added back to the final area totals.

In cases where we have Developed, Low to High Intensity with some tree canopy, the pixel still gets assigned "Urban_Tree" and the area is recomputed, but this time, the residual is labeled as "Built-Up".

This does result in a full accounting of all cover types by county and ctu, but it can lead to some weird interpretations if you're not careful. Because tree canopy is not available for all years, when it's not available we treat all Developed lands as "Built-Up" which leads to a seemingly dramatic reduction in the Built-Up area when tree canopy data is introduced. To account for this, I've added another column called "land_cover_main" which essentially is the parent category when "Urban_Tree" or "Urban_Grassland" are available – here these are simply labeled "Built-up".